hours {chron}	R Documentation

Return Hours, Minutes, or Seconds from a Times Object

Description

Given a chron or times object, extract the hours, minutes or seconds of the time it represents.

Usage

hours(x)
minutes(x)
seconds(x)

Arguments

x
an object inheriting from class "times", or coercible to such via as.chron.

Value

the corresponding time period as an ordered factor.

See Also

chron, dates, times

Examples

x <- chron(dates = c("02/27/92", "02/27/92", "01/14/92", "02/28/92"),
           times = c("23:03:20", "22:29:56", "01:03:30", "18:21:03"))
h <- hours(x)
y <- runif(4)
boxplot(y ~ h)

[Package chron version 2.3-53 Index]