origin {chron}	R Documentation

Chron Origin

Description

Get and set chron origin.

Usage

origin(x)
origin(x) <- value

Arguments

x
a chron object

value
month, day, year vector

Details

origin extracts the origin of its argument. origin<- sets the origin of its argument.

Value

The origin function returns a month, day, year vector.

Note

The default origin for chron objects is January 1, 1970 and it is recommended that that origin be used. Some chron functions may not perform properly with a non-default origin.

See Also

chron

Examples

ch <- as.chron(Sys.Date())
origin(ch)

[Package chron version 2.3-53 Index]